Refactor DatePicker types
